rework interface to only take index logic
